aop: add a model selection bar to place instances,         change placement mode to place new instances when we have a model selection with a single model

+++ aop/probeDatabase.h
@@ -154,9 +154,9 @@ struct ProbeContextToleranceV2 {
 	}
 
 	void setDefault() {
-		occlusion_integerTolerance = 0.25 * OptixProgramInterface::numProbeSamples;
+		occlusion_integerTolerance = int( 0.25 * OptixProgramInterface::numProbeSamples );
 		colorLab_squaredTolerance = 25;
-		distance_tolerance = 0.25;
+		distance_tolerance = 0.25f;
 	}
 };
